Updated local docker config for development
This is an iterative step that start to isolate the needs of
a local docker setup.

* creates an ethereum backed lock
* erc 20 backed lock
* approves the lock to access funds.
