+++ docker/development/deploy-unlock.js
@@ -1,20 +1,13 @@
 /* eslint-disable no-console */
-
 const {
   deploy,
-  getWeb3Provider,
   WalletService,
   Web3Service,
 } = require('@unlock-protocol/unlock-js')
 const Unlock = require('unlock-abi-0-2').Unlock
 const net = require('net')
 const ethers = require('ethers')
-
-var fs = require('fs')
-
-var testErc20Token = JSON.parse(
-  fs.readFileSync('/standup/TestErc20Token.json', 'utf8')
-)
+const TokenDeployer = require('./deploy-locks')
 
 /*
  * This script is meant to be used in dev environment to deploy a version of the Unlock smart
@@ -23,60 +16,15 @@ var testErc20Token = JSON.parse(
 
 const host = process.env.HTTP_PROVIDER || '127.0.0.1'
 const port = 8545
+let providerURL = `http://${host}:${port}`
 
-let provider = new ethers.providers.JsonRpcProvider(`http://${host}:${port}`, {
+let provider = new ethers.providers.JsonRpcProvider(providerURL, {
   chainId: 1984,
 })
 
 let deployedLockAddress
 let pk = '0x08491b7e20566b728ce21a07c88b12ed8b785b3826df93a7baceb21ddacf8b61'
 let recipientAddress = '0xe29ec42f0b620b1c9a716f79a02e9dc5a5f5f98a'
-
-const deployTestERC20Token = async provider => {
-  let wallet = await provider.getSigner(0)
-
-  let factory = new ethers.ContractFactory(
-    testErc20Token.abi,
-    testErc20Token.bytecode,
-    wallet
-  )
-
-  let testERC20
-  try {
-    testERC20 = await factory.deploy({ gasLimit: 6000000 })
-  } catch (e) {
-    console.log(e)
-  }
-
-  await testERC20.deployed()
-  return testERC20.address
-}
-
-const mintForAddress = async (contract, wallet, recipient) => {
-  // let wallet2 = new ethers.Wallet(pk, provider)
-  let wallet2 = new ethers.Wallet(
-    '0xfd8abdd241b9e7679e3ef88f05b31545816d6fbcaf11e86ebd5a57ba281ce229',
-    provider
-  )
-  let contractWSigner = contract.connect(wallet2)
-  let tx = await contractWSigner.mint(recipient, 500, { gasLimit: 6000000 })
-
-  await tx.wait(2)
-  return tx.hash
-}
-
-const approveContract = async (
-  provider,
-  contract,
-  privateKey,
-  contractAddress
-) => {
-  let purchaserWallet = new ethers.Wallet(privateKey, provider)
-  let contractWPurchaser = contract.connect(purchaserWallet)
-  let approvaltx = await contractWPurchaser.approve(contractAddress, 50)
-  await approvaltx.wait(2)
-  return approvaltx.hash
-}
 
 const serverIsUp = (delay, maxAttempts) =>
   new Promise((resolve, reject) => {
@@ -110,7 +58,7 @@ serverIsUp(1000 /* every second */, 120 /* up to 2 minutes */)
       })
 
       const web3 = new Web3Service({
-        readOnlyProvider: `http://${host}:${port}`,
+        readOnlyProvider: providerURL,
         unlockAddress: newContractInstance.options.address,
       })
 
@@ -131,49 +79,19 @@ serverIsUp(1000 /* every second */, 120 /* up to 2 minutes */)
       })
 
       wallet.on('account.changed', async account => {
-        await wallet.createLock(
-          {
-            expirationDuration: 60 * 5, // 1 minute!
-            keyPrice: '0.01', // 0.01 Eth
-            maxNumberOfKeys: -1, // Unlimited
-          },
-          account
-        )
-
-        let testERC20TokenAddress = await deployTestERC20Token(provider)
-
-        let contract = new ethers.Contract(
-          testERC20TokenAddress,
-          testErc20Token.abi,
-          provider
-        )
-        console.log(`Token Contract Address: ${testERC20TokenAddress}`)
-
-        await new Promise(resolve => {
-          setTimeout(resolve, 1000)
-        })
-
-        let tokenTransferHash = await mintForAddress(
-          contract,
+        TokenDeployer.prepareEnvironment(
           wallet,
+          account,
+          provider,
+          pk,
           recipientAddress
         )
-
-        let approvalTransaction = await approveContract(
-          provider,
-          contract,
-          pk,
-          testERC20TokenAddress
-        )
-
-        console.log(`Token Transfer Hash: ${tokenTransferHash}`)
-        console.log(`Contract Approval Transaction: ${approvalTransaction}`)
       })
 
       wallet.on('network.changed', () => {
         wallet.getAccount()
       })
-      wallet.connect(`http://${host}:${port}`)
+      wallet.connect(providerURL)
     })
   })
   .catch(error => {
