+++ docker/development/package.json
@@ -5,7 +5,7 @@
   "author": "Unlock Inc",
   "license": "ISC",
   "dependencies": {
-    "@unlock-protocol/unlock-js": "0.0.24",
+    "@unlock-protocol/unlock-js": "0.0.29",
     "ethers": "^4.0.26",
     "unlock-abi-0-2": "1.0.0"
   }
