+++ docker/development/standup.sh
